FXML changes
I've added a button to fill the labels, where the name of the excersises have to come.

Right now the button is in the same FXML file as the labels, but it is supposed to be on the home screen.

At this moment the button wil fill a label with data from a variable in the Controller. The data in this variable is supposed to be linked to a database.